# api.yaml
swagger: '2.0'
info:
  title: Brant Ford India
  description: Welcome to the world of Commercial and Residential Research Real Estate Insights.
  version: 1.0 
servers:
  - url: http://localhost:3001/
    description:  Local server 
  - url: http://15.206.133.34:3002/
    description:  Pre Production server
  - url: http://15.206.133.34:3002/
    description:  Production server
host: localhost:3001
basePath: /v1
securityDefinitions: {}
schemes:
- http
consumes:
- application/json
produces:
- application/json
paths:
...
  /email-verify:
    post:
      summary: Email Verify
      tags:
      - Authentication
      operationId: EmailVerify
      deprecated: false
      produces:
      - application/json; charset=utf-8
      parameters:
      - name: Body
        in: body
        required: true
        description: ''
        schema:
          $ref: '#/definitions/EmailVerifyRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/EmailVerify'
          examples:
            application/json; charset=utf-8:
              msg: Your account has been verified, Please login.
              data:
                name: test
                email: cejino1614a@gmail.com
          headers:
            X-Powered-By:
              type: string
              default: Express
            Content-Length:
              type: string
              default: '110'
            ETag:
              type: string
              default: W/"6e-rBYOx7xuYnQkua5k6RMMYcXBeRo"
            Date:
              type: string
              default: Tue, 28 Feb 2023 07:43:47 GMT
            Connection:
              type: string
              default: keep-alive
            Keep-Alive:
              type: string
              default: timeout=5
...
definitions:
...
  EmailVerifyRequest:
    title: EmailVerifyRequest
    example:
      code: sfx0zs7U2qSr~oX
    type: object
    properties:
      code:
        type: string
    required:
    - code
  EmailVerify:
    title: EmailVerify
    example:
      msg: Your account has been verified, Please login.
      data:
        name: test
        email: cejino1614a@gmail.com
    type: object
    properties:
      msg:
        type: string
      data:
        $ref: '#/definitions/Data2'
    required:
    - msg
    - data
  Data2:
    title: Data2
    example:
      name: test
      email: cejino1614a@gmail.com
    type: object
    properties:
      name:
        type: string
      email:
        type: string
    required:
    - name
    - email
...